# swagger.yml
openapi: 3.1.0
info:
  title: Receipt Wrangler API.
  termsOfService:
    there are no TOS at this moment, use at your own risk we take no
    responsibility
  license:
    name: GPLV3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 5.0.0
servers:
  - url: /api
security:
  - bearerAuth: [ ]
paths:
...
  /category/getPagedCategories:
    post:
      tags:
        - Category
      summary: Get paged categories
      description: This will return paged categories
      operationId: getPagedCategories
      requestBody:
        description: Paging and sorting data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagedRequestCommand"
      responses:
        200:
          description: Paged categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedData"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
  /category/:
    get:
      tags:
        - Category
      summary: Get all categories
      description: This will return all categories in the system
      operationId: getAllCategories
      responses:
        200:
          description: All categories in the system
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - Category
      summary: Create category
      description: This will create a category
      operationId: createCategory
      requestBody:
        description: Category to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        200:
          description: The created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
  /category/{categoryId}:
    parameters:
      - in: path
        name: categoryId
        schema:
          type: integer
        required: true
        description: Category Id to get
    put:
      tags:
        - Category
      summary: Update category
      description: This will update a category
      operationId: updateCategory
      requestBody:
        description: Category to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        200:
          $ref: "#/components/responses/Ok"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - Category
      summary: Delete category
      description: This will delete a category by id
      operationId: deleteCategory
      responses:
        200:
          $ref: "#/components/responses/Ok"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
  /comment/:
    post:
      tags:
        - Comment
      summary: Add comment
      description: This will add a comment to a receipt, [SYSTEM USER]
      requestBody:
        description: Comment to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCommentCommand"
      operationId: addComment
      responses:
        200:
          description: The created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
  /comment/{commentId}:
    delete:
      tags:
        - Comment
      summary: Delete comment
      description: This will delete a comment by id [SYSTEM User]
      parameters:
        - in: path
          name: commentId
          schema:
            type: integer
          required: true
          description: Comment Id to delete
      operationId: deleteComment
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
...
  /prompt/:
    post:
      tags:
        - Prompt
      summary: Create prompt
      description: This will create a prompt
      operationId: createPrompt
      requestBody:
        description: Prompt to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertPromptCommand"
      responses:
        200:
          description: The created prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
  /prompt/{id}:
    get:
      tags:
        - Prompt
      summary: Get prompt by id
      description: This will get a prompt by id
      operationId: getPromptById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Id of prompt to get
      responses:
        200:
          description: The prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
    put:
      tags:
        - Prompt
      summary: Update prompt by id
      description: This will update a prompt by id
      operationId: updatePromptById
      requestBody:
        description: Prompt to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertPromptCommand"
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Id of prompt to update
      responses:
        200:
          description: The updated prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
    delete:
      tags:
        - Prompt
      summary: Delete prompt by id
      description: This will delete a prompt by id
      operationId: deletePromptById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Id of prompt to delete
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
  /prompt/getPagedPrompts:
    post:
      tags:
        - Prompt
      summary: Gets paged prompts
      description: This will return paged prompts
      operationId: getPagedPrompts
      requestBody:
        description: Paging and sorting data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagedRequestCommand"
      responses:
        200:
          description: Paged prompts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedData"
        500:
          $ref: "#/components/responses/Internal"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
      security:
        - bearerAuth: [ ]
  /prompt/createDefaultPrompt:
    post:
      tags:
        - Prompt
      summary: Create default prompt
      description: This will create a default prompt
      operationId: createDefaultPrompt
      responses:
        200:
          description: The created prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        500:
          $ref: "#/components/responses/Internal"
        403:
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Request was successfully processed
    BadRequest:
      description: The request was malformed
    Forbidden:
      description: The request was not allowed
    Internal:
      description: There was an error processing the request
  schemas:
    ReceiptStatus:
      description: Status of a receipt
      type: string
      enum:
        - "OPEN"
        - "NEEDS_ATTENTION"
        - "RESOLVED"
        - "DRAFT"
        - ""
...
    GroupStatus:
      type: string
      enum:
        - "ACTIVE"
        - "ARCHIVED"
...
    SortDirection:
      type: string
      enum:
        - asc
        - desc
        - ""
...
    SystemTaskStatus:
      type: string
      enum:
        - "SUCCEEDED"
        - "FAILED"
...
    AssociatedEntityType:
      type: string
      enum:
        - "NOOP_ENTITY_TYPE"
        - "RECEIPT"
        - "SYSTEM_EMAIL"
        - "RECEIPT_PROCESSING_SETTINGS"
        - "PROMPT"
...
    BaseModel:
      type: object
      required:
        - id
        - createdAt
      properties:
        id:
          type: integer
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
          default: 0
        createdByString:
          type: string
          description: Created by entity's name
          default: ""
        updatedAt:
          type: string
          default: ""
    Category:
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Name of the category
        description:
          type: string
          description: Description of the category
        updatedAt:
          type: string
          x-go-name: UpdatedAt
      description: Category to relate receipts to
...
    Comment:
      required:
        - comment
        - receiptId
        - userId
        - id
      type: object
      properties:
        additionalInfo:
          type: string
          description: Additional information about the comment
        comment:
          type: string
          description: Comment itself
        commentId:
          type: integer
          description: Comment foreign key used for repleis
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        receiptId:
          type: integer
          description: Receipt foreign key
        updatedAt:
          type: string
        userId:
          type: integer
          description: User foreign key
          format: uint64
      description: User comment left on receipts
    FileData:
      required:
        - receiptId
        - id
      type: object
      properties:
        createdAt:
          type: string

          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
          x-go-name: CreatedBy
        fileType:
          type: string
          description: MIME file type
          x-go-name: FileType
        id:
          type: integer
          format: uint64
          x-go-name: ID
        imageData:
          type: array
          description: Image data
          items:
            type: integer
            format: uint8
          x-go-name: ImageData
        name:
          type: string
          description: File name
          x-go-name: Name
        receiptId:
          type: integer
          description: Receipt foreign key
          format: uint64
          x-go-name: ReceiptId
        size:
          type: integer
          description: File size
          format: uint64
          x-go-name: Size
        updatedAt:
          type: string

          x-go-name: UpdatedAt
      description: File data for images on a receipt
      x-go-package: receipt-wrangler/api/internal/models
    Group:
      required:
        - name
        - status
        - id
        - groupMembers
        - isAllGroup
        - groupReceiptSettings
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        groupSettings:
          $ref: "#/components/schemas/GroupSettings"
        groupReceiptSettings:
          $ref: "#/components/schemas/GroupReceiptSettings"
        groupMembers:
          type: array
          description: Members of the group
          items:
            $ref: "#/components/schemas/GroupMember"
        id:
          type: integer
          format: uint64
        isDefault:
          type: boolean
          description: Is default group (not used yet)
        name:
          type: string
          description: Name of the group
        isAllGroup:
          type: boolean
          description: Is all group for user
        status:
          type: string
          $ref: "#/components/schemas/GroupStatus"
        updatedAt:
          type: string
      description: Group in the system
...
    Receipt:
      required:
        - amount
        - date
        - groupId
        - name
        - paidByUserId
        - status
        - id
      type: object
      properties:
        amount:
          type: string
          description: Receipt total amount
        categories:
          type: array
          description: Categories associated to receipt
          items:
            $ref: "#/components/schemas/Category"
        comments:
          type: array
          description: Comments associated to receipt
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        date:
          type: string
          description: Receipt date
        groupId:
          type: integer
          description: Group foreign key
          format: uint64
        id:
          type: integer
          format: uint64
        imageFiles:
          type: array
          description: Files associated to receipt
          items:
            $ref: "#/components/schemas/FileData"
        name:
          type: string
          description: Receipt name
        paidByUserId:
          type: integer
          description: User paid foreign key
          format: uint64
        receiptItems:
          type: array
          description: Items associated to receipt
          items:
            $ref: "#/components/schemas/Item"
        resolvedDate:
          type: string
          description: Date resolved
        status:
          $ref: "#/components/schemas/ReceiptStatus"
        tags:
          type: array
          description: Tags associated to receipt
          items:
            $ref: "#/components/schemas/Tag"
        updatedAt:
          type: string
        createdByString:
          type: string
          description: Created by string, which is anything that is not a user
      description: Receipt
    Tag:
      required:
        - name
      type: object
      properties:
        createdAt:
          type: string
          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Tag name
        description:
          type: string
          description: Tag description
        updatedAt:
          type: string
      description: Tag to relate receipts to
...
    TagView:
      required:
        - id
        - name
        - numberOfReceipts
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Name of the tag
        description:
          type: string
          description: Description of the tag
        updatedAt:
          type: string
        numberOfReceipts:
          type: integer
          description: Number of receipts associated with this tag
      description: Tag to relate receipts to
...
    PagedRequestCommand:
      required:
        - page
        - pageSize
      type: object
      properties:
        page:
          type: integer
          description: Page number
        pageSize:
          type: integer
          description: Number of records per page
        orderBy:
          type: string
          description: field to order on
        sortDirection:
          type: string
          $ref: "#/components/schemas/SortDirection"
...
    PagedData:
      required:
        - data
        - totalCount
      type: object
      properties:
        data:
          type: array
          items:
            anyOf:
              - $ref: "#/components/schemas/Receipt"
              - $ref: "#/components/schemas/Category"
              - $ref: "#/components/schemas/Tag"
              - $ref: "#/components/schemas/Prompt"
              - $ref: "#/components/schemas/Group"
              - $ref: "#/components/schemas/TagView"
              - $ref: "#/components/schemas/SystemTask"
              - $ref: "#/components/schemas/ReceiptProcessingSettings"
              - $ref: "#/components/schemas/SystemEmail"
        totalCount:
          type: integer
...
    UpsertCommentCommand:
      type: object
      required:
        - comment
        - receiptId
      properties:
        comment:
          type: string
          description: Comment itself
        receiptId:
          type: integer
          description: Receipt foreign key
        userId:
          type: integer
          description: User foreign key
    SystemEmail:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            host:
              type: string
              description: IMAP host
            port:
              type: string
              description: IMAP port
            username:
              type: string
              description: IMAP username
            password:
              type: string
              description: IMAP password
...
    SystemTask:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            type:
              $ref: "#/components/schemas/SystemTaskType"
            status:
              $ref: "#/components/schemas/SystemTaskStatus"
            startedAt:
              type: string
            endedAt:
              type: string
            associatedEntityId:
              type: integer
            associatedEntityType:
              $ref: "#/components/schemas/AssociatedEntityType"
            ranByUserId:
              type: integer
              format: uint64
            resultDescription:
              type: string
            childSystemTasks:
              type: array
              items:
                $ref: "#/components/schemas/SystemTask"
...
    ReceiptProcessingSettings:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            name:
              type: string
              description: Name of the settings
            description:
              type: string
              description: Description of the settings
            aiType:
              $ref: "#/components/schemas/AiType"
            url:
              type: string
              description: URL for custom endpoints
            key:
              type: string
              description: Key for endpoints that require authentication
            model:
              type: string
              description: LLM model
            isVisionModel:
              type: boolean
              description: Is vision model
            ocrEngine:
              $ref: "#/components/schemas/OcrEngine"
            prompt:
              $ref: "#/components/schemas/Prompt"
            promptId:
              type: integer
              description: Prompt foreign key
...
    Prompt:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - name
            - prompt
          properties:
            name:
              type: string
              description: Prompt name
            description:
              type: string
              description: Prompt description
            prompt:
              type: string
              description: Prompt text
    UpsertPromptCommand:
      type: object
      required:
        - name
        - prompt
      properties:
        name:
          type: string
          description: Prompt name
        description:
          type: string
          description: Prompt description
        prompt:
          type: string
          description: Prompt text
...